add the internet check routin for go mod vendor.
Former-commit-id: c2aa86a69726b816eded0150f8a20fbd5dc768bc

+++ DAgent/localbuild.sh
@@ -1,7 +1,14 @@
 #!/bin/bash
 
-rm -rf vendor/A-module
-cp -rf ../A-module ./vendor/
+wget -q --tries=1 --timeout=3 --spider http://google.com
+
+if [[ $? -eq 0 ]]; then
+	echo "Online"
+else
+	echo "Offline"
+	rm -rf vendor/A-module
+	cp -rf ../A-module ./vendor/
+fi
 
 go build -mod vendor
 mv DAgent ./bin
